Docker_project/etl_job/requirements.txt
pymongo==3.12.0
SQLAlchemy==1.4.15
psycopg2-binary==2.9.3
pandas
numpy
vaderSentiment